refacto(api): move createUserAndReconcileToOrganizationLearnerFromExternalUser usecase to right context
